# photosi.api/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PhotoSi.Catalog",
    "version": "v1"
  },
  "paths": {
    "/api/Categorie": {
      "get": {
        "tags": [
          "Categorie"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categoria"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categoria"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorie"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          }
        }
      }
    },
    "/api/Categorie/{id}": {
      "get": {
        "tags": [
          "Categorie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categorie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Categorie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Prodotti": {
      "get": {
        "tags": [
          "Prodotti"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prodotto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prodotto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prodotti"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdottoActionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdottoActionResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Prodotti/{id}": {
      "get": {
        "tags": [
          "Prodotti"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prodotto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prodotto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Prodotti"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Prodotto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Prodotti"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionResult": {
        "type": "object",
        "additionalProperties": false
      },
      "Categoria": {
        "required": [
          "active",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "active": {
            "maxLength": 1,
            "type": "string",
            "default": "S",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Prodotto": {
        "required": [
          "active",
          "categoriaId",
          "codice",
          "id",
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codice": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "name": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "categoriaId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "active": {
            "maxLength": 1,
            "type": "string",
            "default": "S",
            "nullable": true
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        },
        "additionalProperties": false
      },
      "ProdottoActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "$ref": "#/components/schemas/Prodotto"
          }
        },
        "additionalProperties": false
      }
    }
  }
}